requirements.txt
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
tqdm>=4.66.0
pyarrow>=15.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
networkx>=3.0
